Refactored : Project Structure, removed unused imports, deleted some files
