+++ src/Components/Education/Bachelor.js
@@ -2,7 +2,7 @@ import {Stack} from '@mui/material'
 import React from 'react'
 import LocationOnIcon from '@mui/icons-material/LocationOn';
 import DoneIcon from '@mui/icons-material/Done';
-import { data } from '../../Constant/data';
+import { data } from '../Constant/data';
 
 const Bachelor = () => {
