+++ src/Components/ExtraCards/Motivation.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { Stack } from '@mui/material'
-import { Quote } from '../Data'
 import { data } from '../Constant/data'
 
 const Motivation = () => {
@@ -8,7 +7,7 @@ const Motivation = () => {
     return (
     <Stack direction="column" spacing={2} className="MotivationCard">
         <h2> Quote <span> {data.textIcon.writing} </span> </h2>
-        <p> {data.code} "{Quote}" </p>
+        <p> {data.code} "{data.Quote}" </p>
     </Stack>
   )
 }
