+++ src/Components/ExtraCards/Watch.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import InsertDriveFileIcon from '@mui/icons-material/InsertDriveFile';
 import {Stack} from '@mui/material'
 import AnalogClock from 'analog-clock-react';
