+++ src/Components/Personal/Skills.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import Button from '@mui/material/Button'
 import {Stack} from '@mui/material'
 
 const Skills = () => {
