--- src/Data.js
@@ -1,1 +1,0 @@
-export const Quote = "Don't push yourself too hard; sometimes, a little effort is good enough.";
