--- src/Extra/Svg.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Svg = () => {
-    return (
-        <div className=''>
-            <svg height="100" width="100">
-                <circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"/>
-            </svg>
-        </div>
-    )
-}
-
-export default Svg
